requirements.txt
altair==6.0.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
invoke==2.2.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.13.0
netmiko==4.6.0
ntc_templates==8.1.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
paramiko==4.0.0
pillow==12.0.0
protobuf==6.33.2
pyarrow==22.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
PyNaCl==1.6.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scp==0.15.0
six==1.17.0
smmap==5.0.2
streamlit==1.52.1
tenacity==9.1.2
textfsm==2.1.0
toml==0.10.2
tornado==6.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
